vm_event: Fix regression caused by b701ccc8
When xc_map_foreign_batch got deprecated reinitializing vm_event on a domain where
an event listener was previously active broke as it relied on th the flag
XEN_DOMCTL_PFINFO_XTAB to indicate that the magic page is not in the physmap.
Manually check the gpfn type, add it to the physmap, and only then try to map it.

Signed-off-by: Tamas K Lengyel <tlengyel@novetta.com>